ai_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
scikit-learn==1.3.2
numpy==1.24.3
python-multipart==0.0.6
joblib==1.3.2